check col size
