+++ fe/fe-core/src/main/java/org/apache/doris/datasource/hive/HiveUtil.java
@@ -17,6 +17,8 @@
 
 package org.apache.doris.datasource.hive;
 
+import org.apache.doris.catalog.Column;
+import org.apache.doris.common.Pair;
 import org.apache.doris.common.UserException;
 import org.apache.doris.fs.remote.BrokerFileSystem;
 import org.apache.doris.fs.remote.RemoteFileSystem;
@@ -24,8 +26,13 @@ import org.apache.doris.fs.remote.RemoteFileSystem;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.hadoop.hive.common.FileUtils;
+import org.apache.hadoop.hive.metastore.api.FieldSchema;
 import org.apache.hadoop.hive.metastore.api.Partition;
+import org.apache.hadoop.hive.metastore.api.SerDeInfo;
+import org.apache.hadoop.hive.metastore.api.StorageDescriptor;
+import org.apache.hadoop.hive.metastore.api.Table;
 import org.apache.hadoop.hive.ql.io.SymlinkTextInputFormat;
 import org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat;
 import org.apache.hadoop.mapred.InputFormat;
@@ -36,8 +43,13 @@ import org.apache.hadoop.util.ReflectionUtils;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 /**
@@ -156,5 +168,88 @@ public final class HiveUtil {
             resultBuilder.put(entry.getKey(), partition);
         }
         return resultBuilder.build();
+    }
+
+    public static Table toHiveTable(HiveTableMetadata hiveTable) {
+        Objects.requireNonNull(hiveTable.getDbName(), "Hive database name should be not null");
+        Objects.requireNonNull(hiveTable.getTableName(), "Hive table name should be not null");
+        Table table = new Table();
+        table.setDbName(hiveTable.getDbName());
+        table.setTableName(hiveTable.getTableName());
+        // table.setOwner("");
+        int createTime = (int) System.currentTimeMillis() * 1000;
+        table.setCreateTime(createTime);
+        table.setLastAccessTime(createTime);
+        // table.setRetention(0);
+        String location = hiveTable.getProperties().get(HiveMetadataOps.LOCATION_URI_KEY);
+        Set<String> partitionSet = new HashSet<>(hiveTable.getPartitionKeys());
+        Pair<List<FieldSchema>, List<FieldSchema>> hiveSchema = toHiveSchema(hiveTable.getColumns(), partitionSet);
+
+        table.setSd(toHiveStorageDesc(hiveSchema.first, hiveTable.getBucketCols(), hiveTable.getNumBuckets(),
+                hiveTable.getFileFormat(), location));
+        table.setPartitionKeys(hiveSchema.second);
+
+        // table.setViewOriginalText(hiveTable.getViewSql());
+        // table.setViewExpandedText(hiveTable.getViewSql());
+        table.setTableType("MANAGED_TABLE");
+        table.setParameters(hiveTable.getProperties());
+        return table;
+    }
+
+    private static StorageDescriptor toHiveStorageDesc(List<FieldSchema> columns,
+            List<String> bucketCols, int numBuckets, String fileFormat, String location) {
+        StorageDescriptor sd = new StorageDescriptor();
+        sd.setCols(columns);
+        setFileFormat(fileFormat, sd);
+        if (StringUtils.isNotEmpty(location)) {
+            sd.setLocation(location);
+        }
+        sd.setBucketCols(bucketCols);
+        sd.setNumBuckets(numBuckets);
+        Map<String, String> parameters = new HashMap<>();
+        parameters.put("tag", "doris external hive talbe");
+        sd.setParameters(parameters);
+        return sd;
+    }
+
+    private static void setFileFormat(String fileFormat, StorageDescriptor sd) {
+        String inputFormat;
+        String outputFormat;
+        String serDe;
+        if (fileFormat.equalsIgnoreCase("orc")) {
+            inputFormat = "org.apache.hadoop.hive.ql.io.orc.OrcInputFormat";
+            outputFormat = "org.apache.hadoop.hive.ql.io.orc.OrcOutputFormat";
+            serDe = "org.apache.hadoop.hive.ql.io.orc.OrcSerde";
+        } else if (fileFormat.equalsIgnoreCase("parquet")) {
+            inputFormat = "org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat";
+            outputFormat = "'org.apache.hadoop.hive.ql.io.parquet.MapredParquetOutputFormat";
+            serDe = "org.apache.hadoop.hive.ql.io.parquet.serde.ParquetHiveSerDe";
+        } else {
+            throw new IllegalArgumentException("Creating table with an unsupported file format: " + fileFormat);
+        }
+        SerDeInfo serDeInfo = new SerDeInfo();
+        serDeInfo.setSerializationLib(serDe);
+        sd.setSerdeInfo(serDeInfo);
+        sd.setInputFormat(inputFormat);
+        sd.setOutputFormat(outputFormat);
+    }
+
+    private static Pair<List<FieldSchema>, List<FieldSchema>> toHiveSchema(List<Column> columns,
+            Set<String> partitionSet) {
+        List<FieldSchema> hiveCols = new ArrayList<>();
+        List<FieldSchema> hiveParts = new ArrayList<>();
+        for (Column column : columns) {
+            FieldSchema hiveFieldSchema = new FieldSchema();
+            // TODO: add doc, just support doris type
+            hiveFieldSchema.setType(HiveMetaStoreClientHelper.dorisTypeToHiveType(column.getType()));
+            hiveFieldSchema.setName(column.getName());
+            hiveFieldSchema.setComment(column.getComment());
+            if (partitionSet.contains(column.getName())) {
+                hiveParts.add(hiveFieldSchema);
+            } else {
+                hiveCols.add(hiveFieldSchema);
+            }
+        }
+        return Pair.of(hiveCols, hiveParts);
     }
 }
