+++ fe/fe-core/src/main/java/org/apache/doris/datasource/hive/ThriftHMSCachedClient.java
@@ -18,9 +18,7 @@
 package org.apache.doris.datasource.hive;
 
 import org.apache.doris.analysis.TableName;
-import org.apache.doris.catalog.Column;
 import org.apache.doris.common.Config;
-import org.apache.doris.common.Pair;
 import org.apache.doris.datasource.DatabaseMetadata;
 import org.apache.doris.datasource.TableMetadata;
 import org.apache.doris.datasource.hive.event.MetastoreNotificationFetchException;
@@ -28,7 +26,6 @@ import org.apache.doris.datasource.property.constants.HMSProperties;
 
 import com.aliyun.datalake.metastore.hive2.ProxyMetaStoreClient;
 import com.amazonaws.glue.catalog.metastore.AWSCatalogMetastoreClient;
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
@@ -69,18 +66,14 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.security.PrivilegedExceptionAction;
-import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Queue;
-import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
@@ -179,7 +172,7 @@ public class ThriftHMSCachedClient implements HMSCachedClient {
                 // String location,
                 if (tbl instanceof HiveTableMetadata) {
                     ugiDoAs(() -> {
-                        client.client.createTable(toHiveTable((HiveTableMetadata) tbl));
+                        client.client.createTable(HiveUtil.toHiveTable((HiveTableMetadata) tbl));
                         return null;
                     });
                 }
@@ -190,93 +183,6 @@ public class ThriftHMSCachedClient implements HMSCachedClient {
         } catch (Exception e) {
             throw new HMSClientException("failed to create database from hms client", e);
         }
-    }
-
-    @VisibleForTesting
-    public static Table toHiveTable(HiveTableMetadata hiveTable) {
-        Objects.requireNonNull(hiveTable.getDbName(), "Hive database name should be not null");
-        Objects.requireNonNull(hiveTable.getTableName(), "Hive table name should be not null");
-        Table table = new Table();
-        table.setDbName(hiveTable.getDbName());
-        table.setTableName(hiveTable.getTableName());
-        // table.setOwner("");
-        int createTime = (int) System.currentTimeMillis() * 1000;
-        table.setCreateTime(createTime);
-        table.setLastAccessTime(createTime);
-        // table.setRetention(0);
-        String location = hiveTable.getProperties().get(HiveMetadataOps.LOCATION_URI_KEY);
-        Set<String> partitionSet = new HashSet<>(hiveTable.getPartitionKeys());
-        Pair<List<FieldSchema>, List<FieldSchema>> hiveSchema = toHiveSchema(hiveTable.getColumns(), partitionSet);
-
-        table.setSd(toHiveStorageDesc(hiveSchema.first, hiveTable.getBucketCols(), hiveTable.getNumBuckets(),
-                hiveTable.getFileFormat(), location));
-        table.setPartitionKeys(hiveSchema.second);
-
-        // table.setViewOriginalText(hiveTable.getViewSql());
-        // table.setViewExpandedText(hiveTable.getViewSql());
-        table.setTableType("MANAGED_TABLE");
-        table.setParameters(hiveTable.getProperties());
-        return table;
-    }
-
-    private static StorageDescriptor toHiveStorageDesc(List<FieldSchema> columns,
-                                                       List<String> bucketCols,
-                                                       int numBuckets,
-                                                       String fileFormat,
-                                                       String location) {
-        StorageDescriptor sd = new StorageDescriptor();
-        sd.setCols(columns);
-        setFileFormat(fileFormat, sd);
-        if (StringUtils.isNotEmpty(location)) {
-            sd.setLocation(location);
-        }
-        sd.setBucketCols(bucketCols);
-        sd.setNumBuckets(numBuckets);
-        Map<String, String> parameters = new HashMap<>();
-        parameters.put("tag", "doris external hive talbe");
-        sd.setParameters(parameters);
-        return sd;
-    }
-
-    private static void setFileFormat(String fileFormat, StorageDescriptor sd) {
-        String inputFormat;
-        String outputFormat;
-        String serDe;
-        if (fileFormat.equalsIgnoreCase("orc")) {
-            inputFormat = "org.apache.hadoop.hive.ql.io.orc.OrcInputFormat";
-            outputFormat = "org.apache.hadoop.hive.ql.io.orc.OrcOutputFormat";
-            serDe = "org.apache.hadoop.hive.ql.io.orc.OrcSerde";
-        } else if (fileFormat.equalsIgnoreCase("parquet")) {
-            inputFormat = "org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat";
-            outputFormat = "'org.apache.hadoop.hive.ql.io.parquet.MapredParquetOutputFormat";
-            serDe = "org.apache.hadoop.hive.ql.io.parquet.serde.ParquetHiveSerDe";
-        } else {
-            throw new IllegalArgumentException("Creating table with an unsupported file format: " + fileFormat);
-        }
-        SerDeInfo serDeInfo = new SerDeInfo();
-        serDeInfo.setSerializationLib(serDe);
-        sd.setSerdeInfo(serDeInfo);
-        sd.setInputFormat(inputFormat);
-        sd.setOutputFormat(outputFormat);
-    }
-
-    private static Pair<List<FieldSchema>, List<FieldSchema>> toHiveSchema(List<Column> columns,
-                Set<String> partitionSet) {
-        List<FieldSchema> hiveCols = new ArrayList<>();
-        List<FieldSchema> hiveParts = new ArrayList<>();
-        for (Column column : columns) {
-            FieldSchema hiveFieldSchema = new FieldSchema();
-            // TODO: add doc, just support doris type
-            hiveFieldSchema.setType(HiveMetaStoreClientHelper.dorisTypeToHiveType(column.getType()));
-            hiveFieldSchema.setName(column.getName());
-            hiveFieldSchema.setComment(column.getComment());
-            if (partitionSet.contains(column.getName())) {
-                hiveParts.add(hiveFieldSchema);
-            } else {
-                hiveCols.add(hiveFieldSchema);
-            }
-        }
-        return Pair.of(hiveCols, hiveParts);
     }
 
     @Override
